+++ cmd/root.go
@@ -11,7 +11,7 @@ import (
 	"github.com/zimeg/instant-band-night/cmd/musician"
 	"github.com/zimeg/instant-band-night/internal/errors"
 	"github.com/zimeg/instant-band-night/internal/terminal"
-	"github.com/zimeg/instant-band-night/pkg/event"
+	ibnevent "github.com/zimeg/instant-band-night/pkg/event"
 )
 
 // rootCommandFlagSet contains global flags for this program
@@ -23,9 +23,16 @@ type rootCommandFlagSet struct {
 // rootCommandFlags implements the global flags for this prompts
 var rootCommandFlags rootCommandFlagSet
 
+// rootCommandOptions contains values to persist for the command
+type rootCommandOptionSet struct {
+	configDir string
+	eventDate string
+}
+
 // rootCommandNew creates the top level command
 func rootCommandNew() *cobra.Command {
-	tonight := event.Event{}
+	rootCommandOptions := rootCommandOptionSet{}
+	event := ibnevent.Event{}
 	rootCommand := &cobra.Command{
 		Use:   "ibn",
 		Short: "🎶 Instant Band Night CLI",
@@ -41,11 +48,11 @@ func rootCommandNew() *cobra.Command {
 	rootCommand.CompletionOptions.DisableDefaultCmd = true
 	rootCommand.SilenceErrors = true
 	rootCommand.SilenceUsage = true
-	rootCommand.AddCommand(band.BandCommandNew(&tonight))
-	rootCommand.AddCommand(musician.MusicianCommandNew(&tonight))
+	rootCommand.AddCommand(band.BandCommandNew(&event))
+	rootCommand.AddCommand(musician.MusicianCommandNew(&event))
 	aliases := []*cobra.Command{
-		band.BandCommandCreateNew(&tonight),
-		musician.MusicianCommandJoinNew(&tonight),
+		band.BandCommandCreateNew(&event),
+		musician.MusicianCommandJoinNew(&event),
 	}
 	for _, alias := range aliases {
 		alias.Hidden = true
@@ -54,7 +61,24 @@ func rootCommandNew() *cobra.Command {
 	rootCommand.PersistentFlags().StringVarP(&rootCommandFlags.configFlag, "config", "c", "~/.config/ibn", "path to save data")
 	rootCommand.PersistentFlags().StringVarP(&rootCommandFlags.dateFlag, "date", "d", time.Now().Format("2006-01-02"), "date of the event")
 	cobra.OnInitialize(func() {
-		err := loadConfiguration(rootCommand, &tonight)
+		err := setRootCommandOptions(rootCommand, &rootCommandOptions)
+		if err != nil {
+			terminal.PrintError(err)
+			os.Exit(1)
+		}
+		tonight, err := ibnevent.LoadEvent(
+			rootCommandOptions.configDir,
+			rootCommandOptions.eventDate,
+		)
+		if err != nil {
+			terminal.PrintError(err)
+			os.Exit(1)
+		} else {
+			event = tonight
+		}
+	})
+	cobra.OnFinalize(func() {
+		err := ibnevent.SaveEvent(event)
 		if err != nil {
 			terminal.PrintError(err)
 			os.Exit(1)
@@ -63,27 +87,25 @@ func rootCommandNew() *cobra.Command {
 	return rootCommand
 }
 
-// loadConfiguration updates the event of tonight with saved values and settings
-func loadConfiguration(rootCommand *cobra.Command, tonight *event.Event) error {
-	var config string
+// setRootCommandOptions sets the base configurations for the root command setup
+func setRootCommandOptions(
+	rootCommand *cobra.Command,
+	rootCommandOptions *rootCommandOptionSet,
+) (
+	err error,
+) {
 	if !rootCommand.Flag("config").Changed {
-		homedir, err := os.UserHomeDir()
+		homeDir, err := os.UserHomeDir()
 		if err != nil {
 			return err
 		} else {
-			config = fmt.Sprintf("%s/.config/ibn", homedir)
+			rootCommandOptions.configDir = fmt.Sprintf("%s/.config/ibn", homeDir)
 		}
 	} else {
-		config = rootCommand.Flag("config").Value.String()
+		rootCommandOptions.configDir = rootCommand.Flag("config").Value.String()
 	}
-	date := rootCommand.Flag("date").Value.String()
-	loaded, err := event.LoadEvent(config, date)
-	if err != nil {
-		return err
-	} else {
-		*tonight = loaded
-	}
-	return nil
+	rootCommandOptions.eventDate = rootCommand.Flag("date").Value.String()
+	return
 }
 
 // Execute runs the root command of the program