+++ pkg/band/band.go
@@ -3,12 +3,13 @@ package band
 import (
 	"github.com/zimeg/instant-band-night/internal/errors"
 	"github.com/zimeg/instant-band-night/pkg/bucket"
-	"github.com/zimeg/instant-band-night/pkg/instrument"
+	ibninstrument "github.com/zimeg/instant-band-night/pkg/instrument"
 )
 
 // Band contains the lineup of a performing group
 type Band struct {
-	Name        string         `json:"name,omitempty"`
+	id          int
+	Moniker     string         `json:"moniker"`
 	Instruments bucket.Buckets `json:"instruments"`
 }
 
@@ -33,7 +34,7 @@ func NewBand(arrangement Arrangement, buckets bucket.Buckets) (Band, error) {
 // drawInstrument attempts to draw a unique instrument for the band
 func (b *Band) drawInstrument(
 	buckets bucket.Buckets,
-	ins instrument.Instrument,
+	ins ibninstrument.Instrument,
 ) (
 	musicianID string,
 	err error,
@@ -49,4 +50,19 @@ func (b *Band) drawInstrument(
 		}
 	}
 	return "", errors.ErrNotEnoughMusician
+}
+
+// GetID gets the unique and ordered ID of a band
+func (b *Band) GetID() (id int) {
+	return b.id
+}
+
+// GetMusicianIDs returns the musicians in the band
+func (b *Band) GetMusicianIDs() (musicianIDs []string) {
+	return b.Instruments.GetMusicianIDs()
+}
+
+// SetID sets the unique and ordered ID of a band
+func (b *Band) SetID(id int) {
+	b.id = id
 }