+++ pkg/event/instrument.go
@@ -7,12 +7,21 @@ import (
 )
 
 // InstrumentF formats the list of musicians on a certain instrument
-func (e *Event) InstrumentF(band band.Band, ins instrument.Instrument) []string {
-	var musicians []string
+func (e *Event) InstrumentF(
+	band band.Band,
+	ins instrument.Instrument,
+) (
+	musicianIDs []string,
+	musicians []string,
+) {
 	for _, musicianID := range *band.Instruments.GetInstrument(ins) {
-		musician := e.GetMusician(musicianID)
+		musician, err := e.Musicians.GetMusician(musicianID)
+		if err != nil {
+			continue
+		}
+		musicianIDs = append(musicianIDs, musicianID)
 		tag := ins.InstrumentF("%s %s", display.Secondary(musicianID), musician.Name)
 		musicians = append(musicians, tag)
 	}
-	return musicians
+	return musicianIDs, musicians
 }